requirements2.txt
ansitable==0.9.7
bidict==0.22.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colored==2.2.3
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
distlib==0.3.7
dominate==2.8.0
dynamixel-sdk==3.7.31
filelock==3.12.3
firebase-admin==6.2.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-MQTT==1.1.1
Flask-SocketIO==5.3.6
Flask-Testing==0.8.1
Flask-WTF==1.1.1
fonttools==4.42.1
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.26.0
opencv-python==4.8.0.76
paho-mqtt==1.6.1
pgraph-python==0.6.2
Pillow==10.0.1
progress==1.6
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyserial==3.5
python-engineio==4.7.1
python-socketio==5.9.0
pywin32==305.1
requests==2.31.0
roboticstoolbox-python==1.1.0
rsa==4.9
rtb-data==1.0.1
scipy==1.11.2
simple-websocket==0.10.1
spatialgeometry==1.1.0
spatialmath-python==1.1.8
swift-sim==1.1.0
sympy==1.12
torch==2.0.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.4
visitor==0.1.3
websockets==11.0.3
wsproto==1.2.0
WTForms==3.0.1
